Refactor offence class naming and scope
